fix pre-commit ruff errors - remove unused imports and fix import order
