--- deploy/airflow/dags/mlflow_pipeline_dag.py
@@ -1,266 +1,0 @@
-"""
-Homework 3 MLflow Pipeline DAG with drift detection and branching logic.
-"""
-
-from datetime import datetime, timedelta
-from airflow import DAG
-from airflow.operators.python import PythonOperator, BranchPythonOperator
-from airflow.operators.dummy import DummyOperator
-import pandas as pd
-import mlflow
-import json
-from pathlib import Path
-import sys
-import os
-
-# Add src to path for imports (mounted at /opt/airflow/src)
-if "/opt/airflow/src" not in sys.path:
-    sys.path.append("/opt/airflow/src")
-
-# Import pipeline modules
-try:
-    from data_preprocessing import preprocess_data
-    from feature_engineering import engineer_features
-    from model_training import train_base_models, tune_models, build_ensemble
-    from evaluation import evaluate_models, select_and_save_best
-    from drift_detection import detect_drift
-except ImportError as e:
-    print(f"Import error: {e}")
-
-# Set MLflow URI for DAG context (prefer env var from compose)
-mlflow_uri = os.environ.get("MLFLOW_TRACKING_URI", "http://mlflow:5000")
-mlflow.set_tracking_uri(mlflow_uri)
-
-default_args = {
-    "owner": "airflow",
-    "depends_on_past": False,
-    "start_date": datetime(2025, 8, 18),
-    "email_on_failure": False,
-    "email_on_retry": False,
-    "retries": 1,
-    "retry_delay": timedelta(minutes=5),
-}
-
-dag = DAG(
-    "ml_pipeline_dag",
-    default_args=default_args,
-    description="ML Pipeline with MLflow and drift detection",
-    schedule_interval=timedelta(days=7),  # Weekly schedule
-    catchup=False,
-    tags=["homework3", "mlflow", "drift-detection"],
-)
-
-
-def preprocess_data_task(**context):
-    """Preprocess data and generate drifted datasets."""
-    print("Starting data preprocessing...")
-
-    result = preprocess_data("data/global_student_migration.csv", emit_drifted=True)
-
-    if len(result) == 8:
-        (
-            X_train,
-            X_test,
-            y_train,
-            y_test,
-            X_train_drifted,
-            y_train_drifted,
-            X_test_drifted,
-            y_test_drifted,
-        ) = result
-
-        # Save original test set for drift detection
-        original_test = pd.concat([X_test, y_test], axis=1)
-        original_test.to_csv("data/test.csv", index=False)
-
-        print("Data preprocessing completed with drift generation")
-        return "success"
-    else:
-        X_train, X_test, y_train, y_test = result
-        print("Warning: Drifted data not generated")
-        return "success"
-
-
-def feature_engineering_task(**context):
-    """Feature engineering task."""
-    print("Starting feature engineering...")
-
-    # Load preprocessed data - this is simplified for the DAG
-    # In practice, you'd pass data between tasks using XCom or shared storage
-    X_train, X_test, y_train, y_test = preprocess_data(
-        "data/global_student_migration.csv"
-    )
-    X_train_fe, X_test_fe = engineer_features(X_train, X_test)
-
-    print(f"Feature engineering completed. Shape: {X_train_fe.shape}")
-    return "success"
-
-
-def train_model_task(**context):
-    """Train models with MLflow tracking."""
-    print("Starting model training...")
-
-    # Load and process data
-    X_train, X_test, y_train, y_test = preprocess_data(
-        "data/global_student_migration.csv"
-    )
-    X_train_fe, X_test_fe = engineer_features(X_train, X_test)
-
-    # Train base models
-    train_base_models(X_train_fe, y_train, models_dir="models")
-
-    print("Model training completed")
-    return "success"
-
-
-def evaluate_model_task(**context):
-    """Evaluate models and save results."""
-    print("Starting model evaluation...")
-
-    # Load and process data
-    X_train, X_test, y_train, y_test = preprocess_data(
-        "data/global_student_migration.csv"
-    )
-    X_train_fe, X_test_fe = engineer_features(X_train, X_test)
-
-    # This is simplified - in practice you'd load the trained models
-    # For now, we'll just create a simple evaluation result
-    evaluation_results = {
-        "evaluation_timestamp": pd.Timestamp.now().isoformat(),
-        "models_evaluated": 5,
-        "best_model": "randomforest",
-        "best_accuracy": 0.85,
-    }
-
-    # Save evaluation results
-    os.makedirs("reports", exist_ok=True)
-    with open("reports/evaluation_results.json", "w") as f:
-        json.dump(evaluation_results, f, indent=2)
-
-    print("Model evaluation completed")
-    return "success"
-
-
-def drift_detection_task(**context):
-    """Run drift detection and save results."""
-    print("Starting drift detection...")
-
-    try:
-        # Check if required files exist
-        if Path("data/test.csv").exists() and Path("data/drifted_test.csv").exists():
-            drift_results = detect_drift("data/test.csv", "data/drifted_test.csv")
-            print(
-                f"Drift detection completed. Drift detected: {drift_results['drift_detected']}"
-            )
-            return "success"
-        else:
-            print("Warning: Required files for drift detection not found")
-            # Create minimal drift report for demonstration
-            drift_results = {
-                "drift_detected": True,  # Set to True to demonstrate branching
-                "feature_drifts": {
-                    "gpa_or_score": 0.8,
-                    "test_score": 0.7,
-                    "year_of_enrollment": 0.6,
-                },
-                "overall_drift_score": 0.7,
-            }
-
-            os.makedirs("reports", exist_ok=True)
-            with open("reports/drift_report.json", "w") as f:
-                json.dump(drift_results, f, indent=2)
-
-            return "success"
-    except Exception as e:
-        print(f"Drift detection failed: {e}")
-        return "failed"
-
-
-def branch_on_drift(**context):
-    """Branch based on drift detection results."""
-    print("Checking drift detection results for branching...")
-
-    try:
-        # Read drift results from JSON file
-        with open("reports/drift_report.json", "r") as f:
-            drift_results = json.load(f)
-
-        drift_detected = drift_results.get("drift_detected", False)
-
-        if drift_detected:
-            print("Drift detected - branching to retrain_model")
-            return "retrain_model"
-        else:
-            print("No drift detected - branching to pipeline_complete")
-            return "pipeline_complete"
-
-    except Exception as e:
-        print(f"Error reading drift results: {e}")
-        # Default to pipeline_complete if can't read results
-        return "pipeline_complete"
-
-
-def retrain_model_task(**context):
-    """Retrain model with original (non-drifted) data."""
-    print("Starting model retraining due to drift detection...")
-
-    # Load original data (non-drifted)
-    X_train, X_test, y_train, y_test = preprocess_data(
-        "data/global_student_migration.csv", emit_drifted=False
-    )
-    X_train_fe, X_test_fe = engineer_features(X_train, X_test)
-
-    # Retrain models
-    train_base_models(X_train_fe, y_train, models_dir="models")
-
-    print("Model retraining completed")
-    return "success"
-
-
-def pipeline_complete_task(**context):
-    """Simple completion task."""
-    print("Pipeline completed successfully without requiring retraining")
-    return "success"
-
-
-# Define tasks
-preprocess_task = PythonOperator(
-    task_id="preprocess_data", python_callable=preprocess_data_task, dag=dag
-)
-
-feature_eng_task = PythonOperator(
-    task_id="feature_engineering", python_callable=feature_engineering_task, dag=dag
-)
-
-train_task = PythonOperator(
-    task_id="train_model", python_callable=train_model_task, dag=dag
-)
-
-evaluate_task = PythonOperator(
-    task_id="evaluate_model", python_callable=evaluate_model_task, dag=dag
-)
-
-drift_task = PythonOperator(
-    task_id="drift_detection", python_callable=drift_detection_task, dag=dag
-)
-
-branch_task = BranchPythonOperator(
-    task_id="branch_on_drift", python_callable=branch_on_drift, dag=dag
-)
-
-retrain_task = PythonOperator(
-    task_id="retrain_model", python_callable=retrain_model_task, dag=dag
-)
-
-complete_task = DummyOperator(task_id="pipeline_complete", dag=dag)
-
-# Set task dependencies as specified in homework
-(
-    preprocess_task
-    >> feature_eng_task
-    >> train_task
-    >> evaluate_task
-    >> drift_task
-    >> branch_task
-    >> [retrain_task, complete_task]
-)
