+++ deploy/airflow/requirements.txt
@@ -5,5 +5,5 @@ scikit-learn==1.7.1
 joblib==1.5.1
 seaborn==0.13.2
 matplotlib==3.10.3
-mlflow>=2.8.0
-evidently>=0.4.0
+mlflow==3.1.4
+evidently==0.7.11
