+++ requirements.txt
@@ -3,3 +3,8 @@ scikit-learn==1.7.1
 joblib==1.5.1
 seaborn==0.13.2
 matplotlib==3.10.3
+mlflow==3.1.4
+evidently==0.7.11
+psycopg2-binary==2.9.10
+numpy>=1.24.0
+plotly>=5.0.0
